requirements.txt
absl-py==1.0.0
chex==0.1.0
jax>=0.1.55
jaxlib==0.3.0
jaxline==0.0.6
ml-collections==0.1.1
numpy==1.22.2
scipy==1.8.0
tensorflow==2.8.0
chex==0.1.0
dm-haiku==0.0.6
numpy==1.22.2
optax==0.1.1
rlax==0.1.1
editdistance==0.6.0
dm-tree==0.1.6
tensorflow-datasets==4.5.2
